update shell script

+++ run.sh
@@ -38,25 +38,36 @@ while getopts "u:i:p:e:h" opt; do
         ;;
     esac
 done
+
 [ $IP == "" ] && usage
 [ $PORT == "" ] && usage
 [ $USER == "" ] && usage
 [ $PASSWORD == "" ] && usage
 
-echo "$IP $PORT $USER $PASSWORD"
+echo "====== Connection Test (input password) ========"
+ssh -P $PORT $USER@$IP
+status=$(ssh -o ConnectTimeout=5 -p $PORT $USER@$IP echo ok 2>&1)
 
-NAME="${IP}_${PORT}"
-echo "$NAME"
-KEY_NAME="${IP}_${PORT}"
-KEY_PATH="$HOME/.ssh/$KEY_NAME"
-echo "$NAME $KEY_NAME $KEY_PATH" 
-ssh-keygen -t rsa -f $KEY_PATH -N ""
+if [[ $status == ok ]] ; then
+    NAME="${IP}_${PORT}"
+    KEY_NAME="${IP}_${PORT}"
+    KEY_PATH="$HOME/.ssh/$KEY_NAME"
+    ssh-keygen -t rsa -f $KEY_PATH -N ""
 
-echo "Host $NAME" >> ~/.ssh/config
-echo "  HostName $IP" >> ~/.ssh/config
-echo "  User $USER" >> ~/.ssh/config
-echo "  Port $PORT" >> ~/.ssh/config
-echo "  IdentityFile $KEY_PATH" >> ~/.ssh/config
+    echo "" >> ~/.ssh/config
+    echo "Host $NAME" >> ~/.ssh/config
+    echo "  HostName $IP" >> ~/.ssh/config
+    echo "  User $USER" >> ~/.ssh/config
+    echo "  Port $PORT" >> ~/.ssh/config
+    echo "  IdentityFile $KEY_PATH" >> ~/.ssh/config
 
-scp -P $PORT $KEY_PATH.pub $USER@$IP:~/.ssh/$KEY_NAME.pub
-ssh -p $PORT $USER@$IP "cat ~/.ssh/$KEY_NAME.pub >> ~/.ssh/authorized_keys && rm -rf ~/.ssh/$KEY_NAME.pub"
+    scp -P $PORT $KEY_PATH.pub $USER@$IP:~/.ssh/$KEY_NAME.pub
+    ssh -p $PORT $USER@$IP "cat ~/.ssh/$KEY_NAME.pub >> ~/.ssh/authorized_keys && rm -rf ~/.ssh/$KEY_NAME.pub"
+
+elif [[ $status == "Permission denied"* ]] ; then
+  echo no_auth
+else
+  echo other_error
+fi
+
+
